Production ready: DECIMAL precision fixes, migrations, and validation script

+++ backend/models/OrderItem.mjs
@@ -33,11 +33,11 @@ OrderItem.init(
       defaultValue: 1,
     },
     price: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.DECIMAL(10, 2),
       allowNull: false,
     },
     subtotal: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.DECIMAL(10, 2),
       allowNull: false,
     },
     itemType: {
